requirements.txt
lightning
pandas
numpy
mlflow
scikit-learn
tqdm
modal
transformers
torch
cupy-cuda11x
pyarrow